+++ WebContent/WEB-INF/web.xml
@@ -14,6 +14,8 @@
   <session-config>
     <session-timeout>10</session-timeout>
   </session-config>
+  
+  <?ignore
   <filter>
     <filter-name>CAS Authentication Filter</filter-name>
     <filter-class>org.jasig.cas.client.authentication.Saml11AuthenticationFilter</filter-class>
@@ -62,4 +64,5 @@
     <filter-name>CAS HttpServletRequest Wrapper Filter</filter-name>
     <url-pattern>/*</url-pattern>
   </filter-mapping>
+?>
 </web-app>
