+++ src/controllers/ReportsExcelDownloadServlet.java
@@ -1,4 +1,4 @@
-package com.sarf.excel.controller;
+package controllers;
 
 import java.io.IOException;
 import javax.servlet.ServletException;
@@ -6,7 +6,7 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-import com.sarf.excel.model.ReportsExcelCreator;
+import helpers.ReportsExcelCreator;
 
 public class ReportsExcelDownloadServlet extends HttpServlet {
     private static final long serialVersionUID = 1L;
@@ -17,6 +17,7 @@ public class ReportsExcelDownloadServlet extends HttpServlet {
     protected void doGet(HttpServletRequest request,
        HttpServletResponse response)
        throws ServletException, IOException {
+    	doPost(request, response);
       }
 
     protected void doPost(HttpServletRequest request,
